Render a FailureComponent when unauthorized
With this change we introduce a FailureComponent property with
undefined, null and any react node as possible values to enable the
options of:
1. redirecting if FailureComponent is not set (undefined) - default behavior
2. render nothing when unauthorized when FailureComponent is null
3. render other component when unauthorized when FailureComponent is set to any React node

USE CASE:
This is helpful for example if the user wants to use mostly the same
set of rules for route authorization and rendering the correct menu
items for a role. We can wrap MenuItems as we wrap route components
but in MenuItems we want to render only those accessible instead of
redirecting. Another case could be dashboard widgets or actions per
role.